ashmem: int/size_t and slight modernization.
ashmem_get_size_region() should return size_t, since
ashmem_create_region() takes a size_t.

This is a no-op (other than sign) for LP32, since size_t is 32 bits.

For LP64, this is an ABI change, but in effect it just means that we
don't zero the top 32 bits of the result register; any code that worked
before still works, and (highly unlikely) code that allocated > 2GiB
ashmem regions can now happily query that allocated size.

Also lose bogus "\n"s from logging, and use unique_fd rather than goto.

Bug: http://b/205348471
Test: treehugger

+++ libcutils/ashmem-host.cpp
@@ -85,12 +85,10 @@ int ashmem_unpin_region(int /*fd*/, size_t /*offset*/, size_t /*len*/) {
     return 0 /*ASHMEM_IS_UNPINNED*/;
 }
 
-int ashmem_get_size_region(int fd)
-{
+size_t ashmem_get_size_region(int fd) {
     struct stat buf;
     if (!ashmem_validate_stat(fd, &buf)) {
         return -1;
     }
-
     return buf.st_size;
 }
